layout/base/tests/bug1550869-2d.html

<!DOCTYPE HTML>
<html class="reftest-wait">
  <title>Bug 1516963: Test AccessibleCaret doesn't show when clicking on an empty svg container.</title>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
  #container {
    border: 1px solid blue;
    width: 100px;
    height: 100px;
    background-color: yellow;
    outline: 0;
  }
  </style>
  <script>
  document.addEventListener("selectionchange", () => {
    ok(window.getSelection().isCollapsed, "The selection should be collapsed!");
  });

  function click() {
    let container = document.getElementById("container");
    synthesizeMouseAtCenter(container, {});
    setTimeout(() => { document.documentElement.removeAttribute("class"); });
  }
  </script>
  <body onload="SimpleTest.waitForFocus(click);">
    <svg id="container"></svg>
  </body>
</html>
